1차 화면 연결 종료
