+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,16 @@
         android:theme="@style/Theme.Mtot"
         tools:targetApi="31">
         <activity
-            android:name=".ui.account.AddGroupActivity"
+            android:name=".ui.social.AddGroupActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.account.AddFriendActivity"
+            android:name=".ui.social.AddFriendActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.account.SetProfileActivity"
+            android:name=".ui.account.EditProfileActivity"
+            android:exported="false" />
+        <activity
+            android:name=".SetProfileActivity"
             android:exported="false" />
         <activity
             android:name=".StartActivity"
