+++ app/src/main/java/com/example/mtot/ui/account/AccountFragment.kt
@@ -1,14 +1,12 @@
 package com.example.mtot.ui.account
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
 import com.example.mtot.databinding.FragmentAccountBinding
-import com.example.mtot.databinding.FragmentCalendarBinding
 
 class AccountFragment : Fragment() {
 
@@ -20,6 +18,12 @@ class AccountFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         binding = FragmentAccountBinding.inflate(inflater, container, false)
+
+        binding.imageView3.setOnClickListener {
+            val i = Intent(requireContext(), EditProfileActivity::class.java)
+            startActivity(i)
+        }
+
 
         return binding.root
     }
