--- app/src/main/java/com/example/mtot/ui/account/AddGroupActivity.kt
@@ -1,12 +1,0 @@
-package com.example.mtot.ui.account
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.mtot.R
-
-class AddGroupActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_add_group)
-    }
-}
