+++ app/src/main/java/com/example/mtot/ui/account/EditProfileActivity.kt
@@ -6,31 +6,25 @@ import android.os.Bundle
 import androidx.activity.addCallback
 import com.example.mtot.MainActivity
 import com.example.mtot.StartActivity
+import com.example.mtot.databinding.ActivityEditProfileBinding
 import com.example.mtot.databinding.ActivitySetProfileBinding
 
-class SetProfileActivity : AppCompatActivity() {
-    lateinit var binding : ActivitySetProfileBinding
+class EditProfileActivity : AppCompatActivity() {
+    lateinit var binding : ActivityEditProfileBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivitySetProfileBinding.inflate(layoutInflater)
+        binding = ActivityEditProfileBinding.inflate(layoutInflater)
         setContentView(binding.root)
         init()
     }
     fun init(){
 
         val callback = onBackPressedDispatcher.addCallback {
-            val i = Intent(this@SetProfileActivity, StartActivity::class.java)
-            startActivity(i)
-        }
-
-        binding.textEnd.setOnClickListener {
-            val i = Intent(this@SetProfileActivity, MainActivity::class.java)
-            startActivity(i)
+            finish()
         }
 
         binding.buttonNext.setOnClickListener {
-            val i = Intent(this@SetProfileActivity, StartActivity::class.java)
-            startActivity(i)
+            finish()
         }
 
     }
