+++ app/src/main/res/drawable/ic_account_delete_user.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M11.083,10.184C11.528,9.799 11.885,9.323 12.129,8.788C12.373,8.253 12.5,7.672 12.5,7.084C12.5,5.979 12.061,4.919 11.28,4.137C10.498,3.356 9.438,2.917 8.333,2.917C7.228,2.917 6.168,3.356 5.387,4.137C4.606,4.919 4.167,5.979 4.167,7.084C4.167,7.672 4.293,8.253 4.538,8.788C4.782,9.323 5.139,9.799 5.583,10.184C4.417,10.712 3.427,11.565 2.732,12.641C2.038,13.717 1.668,14.97 1.667,16.25C1.667,16.471 1.754,16.683 1.911,16.84C2.067,16.996 2.279,17.084 2.5,17.084C2.721,17.084 2.933,16.996 3.089,16.84C3.246,16.683 3.333,16.471 3.333,16.25C3.333,14.924 3.86,13.653 4.798,12.715C5.735,11.777 7.007,11.25 8.333,11.25C9.659,11.25 10.931,11.777 11.869,12.715C12.807,13.653 13.333,14.924 13.333,16.25C13.333,16.471 13.421,16.683 13.577,16.84C13.734,16.996 13.946,17.084 14.167,17.084C14.388,17.084 14.6,16.996 14.756,16.84C14.912,16.683 15,16.471 15,16.25C14.999,14.97 14.629,13.717 13.934,12.641C13.24,11.565 12.25,10.712 11.083,10.184ZM8.333,9.584C7.839,9.584 7.356,9.437 6.944,9.162C6.533,8.888 6.213,8.497 6.024,8.04C5.834,7.584 5.785,7.081 5.881,6.596C5.978,6.111 6.216,5.666 6.566,5.316C6.915,4.966 7.361,4.728 7.846,4.632C8.331,4.535 8.833,4.585 9.29,4.774C9.747,4.963 10.137,5.284 10.412,5.695C10.687,6.106 10.833,6.589 10.833,7.084C10.833,7.747 10.57,8.383 10.101,8.851C9.632,9.32 8.996,9.584 8.333,9.584ZM17.425,10L18.092,9.342C18.169,9.264 18.231,9.172 18.273,9.071C18.315,8.969 18.337,8.86 18.337,8.75C18.337,8.64 18.315,8.532 18.273,8.43C18.231,8.329 18.169,8.236 18.092,8.159C18.014,8.081 17.922,8.019 17.82,7.977C17.719,7.935 17.61,7.914 17.5,7.914C17.39,7.914 17.281,7.935 17.18,7.977C17.078,8.019 16.986,8.081 16.908,8.159L16.25,8.825L15.592,8.159C15.514,8.081 15.422,8.019 15.32,7.977C15.219,7.935 15.11,7.914 15,7.914C14.778,7.914 14.565,8.002 14.408,8.159C14.331,8.236 14.269,8.329 14.227,8.43C14.185,8.532 14.163,8.64 14.163,8.75C14.163,8.972 14.251,9.185 14.408,9.342L15.075,10L14.408,10.659C14.33,10.736 14.268,10.828 14.226,10.93C14.184,11.031 14.162,11.14 14.162,11.25C14.162,11.36 14.184,11.469 14.226,11.571C14.268,11.672 14.33,11.764 14.408,11.842C14.486,11.92 14.578,11.982 14.679,12.024C14.781,12.067 14.89,12.089 15,12.089C15.11,12.089 15.219,12.067 15.321,12.024C15.422,11.982 15.514,11.92 15.592,11.842L16.25,11.175L16.908,11.842C16.986,11.92 17.078,11.982 17.18,12.024C17.281,12.067 17.39,12.089 17.5,12.089C17.61,12.089 17.719,12.067 17.82,12.024C17.922,11.982 18.014,11.92 18.092,11.842C18.17,11.764 18.232,11.672 18.274,11.571C18.316,11.469 18.338,11.36 18.338,11.25C18.338,11.14 18.316,11.031 18.274,10.93C18.232,10.828 18.17,10.736 18.092,10.659L17.425,10Z"
+      android:fillColor="#000000"/>
+</vector>
