+++ app/src/main/res/drawable/ic_account_edit.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M18.333,6.034C18.334,5.924 18.313,5.815 18.271,5.714C18.23,5.612 18.169,5.52 18.092,5.442L14.558,1.909C14.48,1.831 14.388,1.77 14.287,1.729C14.185,1.687 14.076,1.666 13.967,1.667C13.857,1.666 13.748,1.687 13.647,1.729C13.545,1.77 13.453,1.831 13.375,1.909L11.017,4.267L1.908,13.375C1.831,13.453 1.77,13.546 1.729,13.647C1.687,13.749 1.666,13.857 1.667,13.967V17.5C1.667,17.721 1.754,17.933 1.911,18.09C2.067,18.246 2.279,18.334 2.5,18.334H6.033C6.15,18.34 6.267,18.322 6.376,18.28C6.485,18.239 6.584,18.174 6.667,18.092L15.725,8.984L18.092,6.667C18.168,6.586 18.23,6.493 18.275,6.392C18.283,6.326 18.283,6.258 18.275,6.192C18.279,6.153 18.279,6.114 18.275,6.075L18.333,6.034ZM5.692,16.667H3.333V14.309L11.608,6.034L13.967,8.392L5.692,16.667ZM15.142,7.217L12.783,4.859L13.967,3.684L16.317,6.034L15.142,7.217Z"
+      android:fillColor="#000000"/>
+</vector>
