+++ app/src/main/res/drawable/ic_account_logout.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M10.492,10.834L8.575,12.742C8.497,12.819 8.435,12.912 8.393,13.013C8.35,13.115 8.329,13.224 8.329,13.334C8.329,13.444 8.35,13.553 8.393,13.654C8.435,13.756 8.497,13.848 8.575,13.925C8.652,14.003 8.745,14.066 8.846,14.108C8.948,14.15 9.057,14.172 9.167,14.172C9.277,14.172 9.386,14.15 9.487,14.108C9.589,14.066 9.681,14.003 9.758,13.925L13.092,10.592C13.167,10.513 13.227,10.419 13.267,10.317C13.35,10.114 13.35,9.887 13.267,9.684C13.227,9.581 13.167,9.488 13.092,9.409L9.758,6.075C9.681,5.998 9.588,5.936 9.487,5.894C9.385,5.852 9.277,5.83 9.167,5.83C9.057,5.83 8.948,5.852 8.846,5.894C8.745,5.936 8.653,5.998 8.575,6.075C8.497,6.153 8.436,6.245 8.394,6.347C8.352,6.448 8.33,6.557 8.33,6.667C8.33,6.777 8.352,6.886 8.394,6.987C8.436,7.089 8.497,7.181 8.575,7.259L10.492,9.167H2.5C2.279,9.167 2.067,9.255 1.911,9.411C1.754,9.567 1.667,9.779 1.667,10C1.667,10.221 1.754,10.433 1.911,10.59C2.067,10.746 2.279,10.834 2.5,10.834H10.492ZM10,1.667C8.443,1.66 6.914,2.09 5.589,2.907C4.263,3.725 3.193,4.897 2.5,6.292C2.401,6.491 2.384,6.721 2.455,6.932C2.525,7.143 2.676,7.318 2.875,7.417C3.074,7.517 3.304,7.533 3.515,7.463C3.726,7.392 3.901,7.241 4,7.042C4.527,5.978 5.328,5.074 6.321,4.424C7.314,3.773 8.463,3.4 9.649,3.342C10.835,3.284 12.014,3.544 13.066,4.094C14.118,4.645 15.004,5.466 15.632,6.474C16.26,7.481 16.607,8.638 16.639,9.825C16.67,11.012 16.384,12.185 15.81,13.224C15.236,14.264 14.394,15.13 13.373,15.736C12.352,16.341 11.187,16.663 10,16.667C8.757,16.673 7.538,16.327 6.483,15.671C5.428,15.016 4.579,14.075 4.033,12.959C3.934,12.76 3.759,12.609 3.549,12.538C3.338,12.468 3.107,12.484 2.908,12.584C2.709,12.683 2.558,12.858 2.488,13.069C2.418,13.28 2.434,13.51 2.533,13.709C3.194,15.038 4.198,16.167 5.441,16.979C6.685,17.79 8.122,18.255 9.605,18.325C11.088,18.394 12.563,18.066 13.877,17.375C15.191,16.684 16.296,15.654 17.078,14.392C17.861,13.13 18.292,11.682 18.327,10.198C18.362,8.714 18,7.247 17.278,5.949C16.557,4.652 15.502,3.571 14.222,2.818C12.942,2.065 11.485,1.668 10,1.667Z"
+      android:fillColor="#000000"/>
+</vector>
