+++ app/src/main/res/drawable/shape_annonymous.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <stroke android:color="@color/black" android:width="1dp"/>
+</shape>
