+++ app/src/main/res/layout/activity_add_friend.xml
@@ -1,118 +1,98 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.account.AddFriendActivity">
+    tools:context=".MainActivity"
+    android:orientation="vertical">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/field_next"
+
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="25dp"
-        android:layout_marginEnd="373dp"
-        android:layout_marginStart="17dp"
-        android:layout_marginBottom="678dp"
-        >
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
         <ImageView
-            android:id="@+id/button_next"
-            android:src="@drawable/ic_back_button"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:id="@+id/imageView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
 
-            />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
-    <TextView
-        android:id="@+id/extra_forCommit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="26sp"
-        android:text=""
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-    <TextView
-        android:id="@+id/text_addFriend"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="25dp"
-        android:layout_marginEnd="255dp"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        android:text="친구 추가"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/field_writeName"
-        android:layout_width="365dp"
-        android:layout_height="45dp"
-        android:layout_marginTop="45dp"
-        android:layout_marginStart="0dp"
-        android:layout_marginBottom="728dp"
-
-        app:layout_constraintStart_toStartOf="@+id/text_addFriend"
-        app:layout_constraintTop_toTopOf="@+id/text_addFriend"
-        >
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/field_myID"
-        android:layout_width="80dp"
-        android:layout_height="20dp"
-        android:layout_marginTop="60dp"
-        android:layout_marginBottom="728dp"
-
-        app:layout_constraintStart_toStartOf="@+id/field_writeName"
-        app:layout_constraintTop_toTopOf="@+id/field_writeName"
-        >
+            android:layout_marginTop="33dp"
+            android:layout_marginStart="10dp"
+            android:src="@drawable/back_icon" />
 
         <TextView
-            android:id="@+id/text_myID"
-            android:layout_width="50dp"
+            android:id="@+id/textView"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:textSize="11sp"
+            android:layout_weight="1"
+
+            android:layout_marginTop="30dp"
+            android:layout_marginEnd="250dp"
+            android:textSize="22dp"
+            android:text="친구 추가"
+            android:textColor="#012A4A"
+            android:textStyle="bold"/>
+    </LinearLayout>
+
+    <EditText
+        android:id="@+id/editTextText"
+        android:layout_width="match_parent"
+        android:layout_height="60dp"
+        android:layout_marginTop="15dp"
+        android:layout_marginStart="30dp"
+        android:layout_marginEnd="30dp"
+        android:ems="10"
+        android:inputType="text"
+        android:text="아이디"
+        />
+
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="5dp"
+
+        android:layout_marginBottom="70dp"
+        android:orientation="horizontal">
+
+        <TextView
+            android:id="@+id/textView2"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+
+            android:layout_marginStart="30dp"
+            android:layout_marginTop="10dp"
             android:text="내 아이디"
-
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/field_add"
-        android:layout_width="220dp"
-        android:layout_height="35dp"
-        android:layout_marginTop="80dp"
-        android:layout_marginBottom="728dp"
-        android:layout_marginStart="70dp"
-
-        app:layout_constraintStart_toStartOf="@+id/field_myID"
-        app:layout_constraintTop_toTopOf="@+id/field_myID"
-        >
+            android:textColor="#000000" />
 
         <TextView
-            android:id="@+id/text_add"
-            android:layout_width="50dp"
+            android:id="@+id/textView3"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:textSize="15sp"
-            android:text="추가"
 
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent" />
+            android:layout_marginStart="260dp"
+            android:layout_marginTop="10dp"
+            android:text="vpql!"
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+            android:textColor="#000000"/>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/button"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+
+        android:layout_marginStart="105dp"
+        android:layout_marginEnd="105dp"
+        android:text="추가"
+        android:textColor="#000000"
+        android:textSize="16dp"
+        app:tint="#A9D6E5"/>
+
+</LinearLayout>
