+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -16,16 +16,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-            android:id="@+id/text_setProf"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="15dp"
-            android:text="프로필 설정"
-            android:textSize="25sp"
-            app:layout_constraintStart_toStartOf="@id/button_next"
-            app:layout_constraintTop_toBottomOf="@id/button_next" />
-
 
         <ImageView
             android:id="@+id/field_profImg"
@@ -34,7 +24,7 @@
             android:layout_marginTop="50dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/text_setProf" />
+            app:layout_constraintTop_toBottomOf="@id/button_next" />
 
 
         <EditText
